chore(origin-ui-certificate-view): make generic component for certificate action content
