+++ packages/ui/libs/certificate/view/src/containers/action/DepositAction/DepositAction.styles.ts
