+++ packages/ui/libs/certificate/view/src/containers/action/DepositAction/DepositAction.tsx
@@ -1,0 +1,11 @@
+import React, { FC } from 'react';
+import { CertificateActionContent } from '../../list';
+
+export const DepositAction: FC = () => {
+  return (
+    <CertificateActionContent
+      title="Selected for Deposit"
+      buttonText="Deposit"
+    />
+  );
+};
