+++ packages/ui/libs/certificate/view/src/containers/action/RetireAction/RetireAction.tsx
@@ -1,0 +1,14 @@
+import React, { FC } from 'react';
+import {
+  CertificateActionContent,
+  CertificateActionContentProps,
+} from '../../list';
+
+export const RetireAction: FC = () => {
+  return (
+    <CertificateActionContent
+      title="Selected for Retirement"
+      buttonText="Retire"
+    />
+  );
+};
