+++ packages/ui/libs/certificate/view/src/containers/action/SellAction/SellAction.styles.ts
@@ -1,0 +1,10 @@
+import { makeStyles } from '@material-ui/core';
+
+export const useStyles = makeStyles({
+  totalPrice: {
+    margin: '20px 0',
+    width: '100%',
+    display: 'flex',
+    justifyContent: 'space-between',
+  },
+});
