+++ packages/ui/libs/certificate/view/src/containers/action/WithdrawAction/WithdrawAction.tsx
@@ -1,0 +1,11 @@
+import React, { FC } from 'react';
+import { CertificateActionContent } from '../../list';
+
+export const WithdrawAction: FC = () => {
+  return (
+    <CertificateActionContent
+      title="Selected for Withdrawal"
+      buttonText="Withdraw"
+    />
+  );
+};
