+++ packages/ui/libs/certificate/view/src/containers/index.ts
@@ -1,1 +1,2 @@
 export * from './list';
+export * from './action';
