+++ packages/ui/libs/certificate/view/src/containers/list/CertificateActionContent/index.ts
@@ -1,0 +1,1 @@
+export * from './CertificateActionContent';
