+++ packages/ui/libs/certificate/view/src/containers/list/index.ts
@@ -1,6 +1,3 @@
 export * from './ListItemHeader';
 export * from './ListItemContent';
-export * from './SellAction';
-export * from './WithdrawAction';
-export * from './DepositAction';
-export * from './RetireAction';
+export * from './CertificateActionContent';
